% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_temp.R
\name{pop_temp}
\alias{pop_temp}
\title{pop_temp}
\usage{
pop_temp(data, template, template_file = "n", output_folder = "./",
  output_name_template = "[[i]]", output_extension = ".txt",
  begin_token = "\\\\[\\\\[", end_token = "\\\\]\\\\]")
}
\arguments{
\item{data}{A data frame containing the values to use.}

\item{template}{The character object to use as a template.}

\item{template_file}{The file to use as a template (overwrites \code{template} if set).}

\item{output_name_template}{A template to use when naming the output files (default: row number).}

\item{output_extension}{File extension of the output files (default .txt).}

\item{begin_token}{Token in the template to begin substitution (default [[).}

\item{end_token}{Token to use in the template to end substitution (default ]]).}
}
\description{
Create a set of files by populating a template with a set of values
taken from rows of a data frame.
}
\examples{

}

